Add some helpers for AfterAll and Scope compatibility
