+++ core/src/main/scala-2/net/ruippeixotog/akka/testkit/specs2/AfterAllCompat.scala
@@ -1,0 +1,9 @@
+package net.ruippeixotog.akka.testkit.specs2
+
+import org.specs2.specification.AfterAll
+
+trait AfterAllCompat extends AfterAll {
+  def shutdownTestkit(): Unit
+
+  def afterAll(): Unit = shutdownTestkit()
+}
