+++ core/src/main/scala-3/net/ruippeixotog/akka/testkit/specs2/ScopeCompat.scala
@@ -1,0 +1,5 @@
+package net.ruippeixotog.akka.testkit.specs2
+
+import org.specs2.execute.Scope
+
+trait ScopeCompat extends Scope
